Stop the debug port splitting lines and truncating values
Two fixes to the only window into a running modem.

A zero value no longer ends the line. writeDebugInt(int16_t) special-cased
zero and called writeDebug("0"), which is the public single-argument overload
and appends a newline, so every field after a zero landed on a line of its
own. The sync report suffers most:

  DEBUG6("Mode2RX: valid sync vector", corr, m_dataPtr, n1, n2, errs)

n1 and errs are zero exactly when everything is working, so a healthy modem
produced the most broken looking output. Zero is now handled inline like any
other digit.

Numbers are no longer truncated to 16 bits. corr is a q31_t and was being
passed through an int16_t parameter; it reaches 17973 on a modest signal, so
a stronger one wraps and reports something negative and meaningless. All six
writeDebug overloads and writeDebugInt now take int32_t.

The conversion buffer grows from 10 to 12 bytes to fit -2147483648 and its
terminator, and the negation goes through an unsigned type so the most
negative value does not overflow. The old code would have overflowed on
INT16_MIN too, it just never saw one.

+++ SerialPort.h
@@ -38,12 +38,14 @@ public:
   void writeKISSData(uint8_t type, const uint8_t* data, uint16_t length);
   void writeKISSAck(uint16_t token);
 
+  // int32_t rather than int16_t: several callers pass a q31_t, which a 16 bit
+  // parameter silently wraps.
   void writeDebug(const char* text);
-  void writeDebug(const char* text, int16_t n1);
-  void writeDebug(const char* text, int16_t n1, int16_t n2);
-  void writeDebug(const char* text, int16_t n1, int16_t n2, int16_t n3);
-  void writeDebug(const char* text, int16_t n1, int16_t n2, int16_t n3, int16_t n4);
-  void writeDebug(const char* text, int16_t n1, int16_t n2, int16_t n3, int16_t n4, int16_t n5);
+  void writeDebug(const char* text, int32_t n1);
+  void writeDebug(const char* text, int32_t n1, int32_t n2);
+  void writeDebug(const char* text, int32_t n1, int32_t n2, int32_t n3);
+  void writeDebug(const char* text, int32_t n1, int32_t n2, int32_t n3, int32_t n4);
+  void writeDebug(const char* text, int32_t n1, int32_t n2, int32_t n3, int32_t n4, int32_t n5);
 
 private:
   uint8_t  m_buffer[2000U];
@@ -55,7 +57,7 @@ private:
 
 #if defined(SERIAL_DEBUGGING)
   void writeDebugInt(const char* text);
-  void writeDebugInt(int16_t num);
+  void writeDebugInt(int32_t num);
   void reverse(char* buffer, uint8_t length) const;
 #endif
